drop V1 swap support: one keyspace, one resolver, no chain rebuild for offers
