+++ src/lib/activityHistory.ts
@@ -1,11 +1,10 @@
 import type { Activity } from '@arkade-os/sdk'
 import { isRfqSwapTerminal } from '@arkade-os/swap/protocol'
-import { ASSET_SWAP_ACTIVITY_KIND } from './swapRecords'
+import { ASSET_SWAP_ACTIVITY_KIND, type LnSendView } from './swapRecords'
 import { consoleError } from './logs'
 import type { TransactionActivityMetadata } from './storage'
 import { buildAssetSwapActivityTx } from './swapDisplay'
 import type { ExitRecord } from './exitHistory'
-import type { LnSendView } from './lnSendRecords'
 import type { WalletAssetSwap } from './swapRepository'
 import { arkTransactionToTx, sortLocalTxs, txidOfArkTransaction } from './transactionHistory'
 import type { Tx } from './types'
